+++ Sources/MapboxMaps/Offline/TileStore+MapboxMaps.swift
@@ -78,13 +78,13 @@ extension TileStore {
             return __loadTileRegion(forId: id,
                                     loadOptions: loadOptions,
                                     onProgress: progress,
-                                    onFinished: tileStoreClosureAdapter(for: completion, type: TileRegion.self))
+                                    onFinished: tileStoreClosureAdapter(for: completion, type: TileRegion.self)).asCancelable()
         }
         // Use overloaded version
         else {
             return __loadTileRegion(forId: id,
                                     loadOptions: loadOptions,
-                                    onFinished: tileStoreClosureAdapter(for: completion, type: TileRegion.self))
+                                    onFinished: tileStoreClosureAdapter(for: completion, type: TileRegion.self)).asCancelable()
         }
     }